Goblint's stdout/stderr redirected into output #8 and logging support added
All debug info that Goblint dumps is now redirected so that it is shown in the Output console in vscode. This is done with zt-exec, which has a method to redirect stdout and stderr to System.err.
As the default runCommand method from MagpieBridge's ToolAnalysis class is no longer used due to the need to use zt-exec's ProcessExecutor instead of ProcessBuilder, GoblintAnalysis now implements ServerAnalysis interface instead of ToolAnalysis.
Logging support was added with Log4j. Logs are shown in the Output console as well. The default level is info, but debug level can be used for development. For using debug level, log4j2.xml file must be edited.
On the info level, the command Goblint was executed with is logged as well as some additional info (e.g. magpie server started).
Some additional exception handling was added to prevent reading from a JSON file after the Goblint run fails, as there were exceptions when another file's analysis result was tried to be shown in a different file.