Gobpie uses location ranges provided by Goblint #12
Instead of trying to find where to end a notification by finding the first ';', GobPie now uses line and column end positions from the JSON output provided by Goblint.
Additonally, the Positions of AnalysisResults get the sourcfile information from JSON instead of using the file that the analysis was stared on as sourcefile, because the JSON output has results for several different files, not only the one that trigerred the analysis.